* master: (66 commits)
  Get rid of most .h files in iop/
  masks: better prevention of memory fragmentation
  calculator: do not use integer variable in read_number() to store intermediate values
  noise profile: panasonic dmc-lx7 by deryk lister
  lua: fix post-image-import not working for imgid >255
  sync darktable-9999.ebuild
  cmake: Drop obsolete EXPERIMENTAL and LEGACY flags
  [l10n] Swedish update of .desktop strings
  Italian translation updated
  Updated Dutch translation
  tiff output: tentatively fix compilation with libtiff 3.9 which lacks tmsize_t
  Show pixelpipe flow info when  -d perf
  Updated Danish translation
  tiff output: several fixes
  Update French translation.
  _Keywords got dropped by mistake, restore it
  Make data/darktable.desktop translatable along with all other things that need translation.  To translate, translators don't need to do anything more than they have done so far, it just beomes a few more strings to translate.
  Add German keywords to .desktop file
  Add italian keywords in .desktop
  tiff output: set sample format for 8bit and 16bit too
  ...

Conflicts:
	src/iop/exposure.h

+++ src/common/calculator.c
@@ -62,7 +62,7 @@ typedef struct parser_state_t
 
 static float read_number(parser_state_t *self)
 {
-  unsigned long int integer = 0, fractional = 0, fractional_div = 1;
+  float integer = 0.0f, fractional = 0.0f, fractional_div = 1.0f;
   while(*self->p && *self->p >= '0' && *self->p <= '9')
   {
     integer = integer * 10 + (*self->p) - '0';
@@ -79,7 +79,7 @@ static float read_number(parser_state_t *self)
     }
   }
 
-  return integer + (float)fractional / fractional_div;
+  return integer + fractional / fractional_div;
 }
 
 static token_t * get_token(parser_state_t *self)
